refactor: kill `hax-cli-options[-engine]`&`hax-diagnostics`, intro `hax-types`
