+++ Cargo.toml
@@ -2,9 +2,6 @@
 members = [
      "frontend/exporter",
      "frontend/exporter/options",
-     "frontend/diagnostics",
-     "cli/options",
-     "cli/options/engine",
      "cli/subcommands",
      "cli/driver",
      "test-harness",
@@ -16,15 +13,12 @@ members = [
      "hax-lib-protocol-macros",
      "hax-bounded-integers",
      "engine/names",
-     "engine/names/extract",
+     "engine/names/extract", "hax-types",
 ]
 exclude = ["tests"]
 default-members = [
      "frontend/exporter",
      "frontend/exporter/options",
-     "frontend/diagnostics",
-     "cli/options",
-     "cli/options/engine",
      "cli/subcommands",
      "cli/driver",
      "test-harness",
@@ -81,12 +75,10 @@ annotate-snippets = "0.11"
 # Crates in this repository
 hax-frontend-exporter = { path = "frontend/exporter", version = "=0.1.0-pre.1" }
 hax-adt-into = { path = "frontend/exporter/adt-into", version = "=0.1.0-pre.1" }
-hax-cli-options = { path = "cli/options", version = "=0.1.0-pre.1" }
-hax-cli-options-engine = { path = "cli/options/engine", version = "=0.1.0-pre.1" }
 hax-frontend-exporter-options = { path = "frontend/exporter/options", version = "=0.1.0-pre.1" }
-hax-diagnostics = { path = "frontend/diagnostics", version = "=0.1.0-pre.1" }
 hax-phase-debug-webapp = { path = "engine/utils/phase-debug-webapp", version = "=0.1.0-pre.1" }
 hax-lib-macros-types = { path = "hax-lib-macros/types", version = "=0.1.0-pre.1" }
 hax-lib-macros = { path = "hax-lib-macros", version = "=0.1.0-pre.1" }
 hax-lib = { path = "hax-lib", version = "=0.1.0-pre.1" }
 hax-engine-names = { path = "engine/names", version = "=0.1.0-pre.1" }
+hax-types = { path = "hax-types", version = "=0.1.0-pre.1" }
