+++ PUBLISHING.md
@@ -14,10 +14,8 @@ and `examples`):
 1. `hax-frontend-exporter-options` (`frontend/exporter/options `)
 2. `hax-adt-into` (`frontend/exporter/adt-into`)
 3. `hax-frontend-exporter` (`frontend/exporter`)
-4. `hax-cli-options` (`cli/options`)
-5. `hax-diagnostics` (`frontend/diagnostics`)
-6. `hax-cli-options-engine` (`cli/options/engine`)
-7. `hax-subcommands` (binaries) (`cli/subcommands`)
+4. `hax-types` (`hax-types`)
+5. `hax-subcommands` (binaries) (`cli/subcommands`)
    - `cargo-hax`
    - `hax-export-json-schemas`
    - `hax-pretty-print-diagnostics`
