+++ cli/driver/Cargo.toml
@@ -22,10 +22,8 @@ serde_json.workspace = true
 clap.workspace = true
 colored.workspace = true
 hax-frontend-exporter.workspace = true
-hax-cli-options.workspace = true
-hax-cli-options-engine.workspace = true
+hax-types.workspace = true
 hax-frontend-exporter-options.workspace = true
-hax-diagnostics.workspace = true
 hax-lib-macros-types.workspace = true
 itertools.workspace = true
 which.workspace = true
