+++ cli/driver/src/callbacks_wrapper.rs
@@ -1,4 +1,4 @@
-use hax_cli_options::{Command, ENV_VAR_OPTIONS_FRONTEND};
+use hax_types::cli_options::{Command, Options, ENV_VAR_OPTIONS_FRONTEND};
 
 use rustc_driver::{Callbacks, Compilation};
 use rustc_interface::{interface, Queries};
@@ -8,14 +8,14 @@ use rustc_span::symbol::Symbol;
 /// configuration in the `config` phase of rustc
 pub struct CallbacksWrapper<'a> {
     pub sub: &'a mut (dyn Callbacks + Send + 'a),
-    pub options: hax_cli_options::Options,
+    pub options: Options,
 }
 impl<'a> Callbacks for CallbacksWrapper<'a> {
     fn config(&mut self, config: &mut interface::Config) {
         let options = self.options.clone();
         config.psess_created = Some(Box::new(move |parse_sess| {
             parse_sess.env_depinfo.get_mut().insert((
-                Symbol::intern(hax_cli_options::ENV_VAR_OPTIONS_FRONTEND),
+                Symbol::intern(ENV_VAR_OPTIONS_FRONTEND),
                 Some(Symbol::intern(&serde_json::to_string(&options).unwrap())),
             ));
         }));
