--- cli/options/engine/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "hax-cli-options-engine"
-version.workspace = true
-authors.workspace = true
-license.workspace = true
-homepage.workspace = true
-edition.workspace = true
-repository.workspace = true
-readme.workspace = true
-description = "Defines the data type of the input of the engine. The (OCaml) engine reads JSON on stdin and deserializes the payload according to the JSON Schemas defined by this crate."
-
-[dependencies]
-serde.workspace = true
-serde_json.workspace = true
-schemars.workspace = true
-itertools.workspace = true
-clap.workspace = true
-hax-diagnostics.workspace = true
-hax-cli-options.workspace = true
-hax-frontend-exporter.workspace = true
-hax-frontend-exporter-options.workspace = true
-path-clean = "1.0.1"
-
