+++ engine/names/extract/Cargo.toml
@@ -11,7 +11,6 @@ description = "Helper binary generating an OCaml module"
 
 
 [build-dependencies]
-hax-cli-options-engine.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 cargo_metadata.workspace = true
