--- frontend/diagnostics/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "hax-diagnostics"
-version.workspace = true
-authors.workspace = true
-license.workspace = true
-homepage.workspace = true
-edition.workspace = true
-repository.workspace = true
-readme.workspace = true
-description = "This crate defines types that represent the various errors that can occur in hax."
-
-[package.metadata.rust-analyzer]
-rustc_private=true
-
-[dependencies]
-serde.workspace = true
-schemars.workspace = true
-colored.workspace = true
-annotate-snippets.workspace = true
-hax-frontend-exporter.workspace = true
