+++ hax-types/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "hax-cli-options"
+name = "hax-types"
 version.workspace = true
 authors.workspace = true
 license.workspace = true
@@ -7,13 +7,15 @@ homepage.workspace = true
 edition.workspace = true
 repository.workspace = true
 readme.workspace = true
-build = "build.rs"
-description = "The CLI options read by the `cargo-hax` binary."
 
 [dependencies]
-serde.workspace = true
-serde_json.workspace = true
-schemars.workspace = true
 clap.workspace = true
+hax-frontend-exporter.workspace = true
 hax-frontend-exporter-options.workspace = true
+itertools.workspace = true
 path-clean = "1.0.1"
+schemars.workspace = true
+serde.workspace = true
+colored.workspace = true
+serde_json.workspace = true
+annotate-snippets.workspace = true
