+++ hax-types/src/cli_options.rs
@@ -1,8 +1,7 @@
+use crate::prelude::*;
+
 use clap::{Parser, Subcommand, ValueEnum};
-use schemars::JsonSchema;
-use serde::{Deserialize, Serialize};
 use std::fmt;
-use std::path::{Path, PathBuf};
 
 pub use hax_frontend_exporter_options::*;
