+++ hax-types/src/diagnostics/mod.rs
@@ -1,14 +1,11 @@
-#![feature(rustc_private)]
-
 extern crate rustc_driver;
 extern crate rustc_error_messages;
 extern crate rustc_errors;
 extern crate rustc_session;
 extern crate rustc_span;
 
+use crate::prelude::*;
 use colored::Colorize;
-use schemars::JsonSchema;
-use serde::{Deserialize, Serialize};
 
 pub mod report;
