+++ hax-types/src/diagnostics/report.rs
