+++ hax-types/src/prelude.rs
@@ -1,0 +1,3 @@
+pub use schemars::JsonSchema;
+pub use serde::{Deserialize, Serialize};
+pub use std::path::{Path, PathBuf};
